Slightly Janky Fix to HTTP Response Issue
Fixed the issue where the contents of the http response were being downloaded by the browser. It is a janky solution because now every post request looks like it receives a 500 error (even though the request is getting through...I basically just blocked responses from the server to prevent this error which is really bad practice). In general, I should switch to using AJAX instead of an action through a form.